Move fetchers all the way up to `bleep-model`
